pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "container-secure-compose"
version = "0.1.0"
description = "Privilege-minimised Docker Compose configuration generator"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[project.scripts]
csc = "csc.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["csc*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=csc --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
